+++ CHANGELOG.md
@@ -9,8 +9,10 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Added airlift json source and airlift json transform to vector.toml ([#553]).
+- Added commons structs as well as helper functions for Affinity ([#556]).
 
 [#553]: https://github.com/stackabletech/operator-rs/pull/553
+[#556]: https://github.com/stackabletech/operator-rs/pull/556
 
 ## [0.34.0] - 2023-02-06